# BackEnd/swagger.yaml
openapi: 3.0.0
info:
  title: Simple Inventory API
  description: This is a simple API
  contact:
    email: you@your-company.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
# Added by API Auto Mocking Plugin
- description: SwaggerHub API Auto Mocking
  url: https://virtserver.swaggerhub.com/MatteoLunardon/TRENTunOAPIs_la_vendetta/1.0.0
- url: https://virtserver.swaggerhub.com/JOAOPROGRAMMING/TRENTunO/1.0.0
  description: TRENTunO API Auto Mocking
...
paths:
...
  /parks:
    get:
      tags:
      - Park
      summary: Searches parks
      description: Returns an array with all parks in the system, optionally filtered by name or category.
      parameters:
      - name: name
        in: query
        description: The name of the park to filter by
        required: false
        schema:
          type: string
      - name: category
        in: query
        description: The category of the park to filter by
        required: false
        schema:
          $ref: '#/components/schemas/CategoryEnum'
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parks'
        "400":
          description: Invalid input parameter

    post:
      tags:
      - Park
      description: Adds a new park to the system.
      requestBody:
        description: The park object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Parks'
      responses:
        "201":
          description: Park successfully created
        "400":
          description: Invalid input format
  /parks/{id}:
    get:
      tags:
      - Park
      description: Returns a park in the system with the specified ID.
      parameters:
      - name: id
        in: path
        description: The unique ID of the park
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parks'
        "400":
          description: Invalid input parameter
        "404":
          description: Park not found
  
    delete:
      tags:
      - Park
      description: Deletes a park in the system with the specified ID.
      parameters:
      - name: id
        in: path
        description: The unique ID of the park to delete
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Park successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Park successfully deleted
        "400":
          description: Invalid ID format
        "404":
          description: Park not found
        "500":
          description: Server error
  /events:
    get:
      tags:
      - Event
      description: This API will all the events in the system
      parameters:
      - name: user_id
        in: query
        description: filter events by User ID
        required: false
        schema:
          type: string
      - name: park_id
        in: query
        description: filter events by Park ID
        schema:
          type: string
      - name: event status
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Events'
        "400":
          description: Invalid input format
        "404":
          description: Event not found
        "500":
          description: Server error
  
    post:
      tags:
      - Event
      description: This API will add a new event by passing the required paramters ( only if the user is authorized to do so)
      parameters:
        - name: UserAuth
          in: query
          description: User Authorization
          required: true 
          schema:
            type: boolean
      requestBody:
        description: The User object to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Events'
      responses:
        "201":
          description: event created
        "400":
          description: invalid input, object invalid
        "500":
          description: Server error
  /events/{id}:
    get:
      tags:
      - Event
      description: This API will return an event in the system which has the id specified
      parameters:
      - name: id
        in: path
        description: Event id
        required: true 
        schema: 
          $ref: '#/components/schemas/Parks'
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Events'
        "400":
          description: Invalid input format
        "404":
          description: Event not found.
        "500":
          description: Server error
    put:
      tags:
        - Event
      summary: Update an event's description
      description: Allows an admin to update the description of an event.
      parameters:
        - name: id
          in: path
          description: The unique ID of the event to update
          required: true
          schema:
            type: string
        - name: user_level
          in: query
          description: The user level to validate admin access
          required: true
          schema:
            type: string
            enum: [Admin, User]
      requestBody:
        description: The updated event description
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  example: "Updated description for the event."
      responses:
        "200":
          description: Event successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Event successfully updated."
                  event:
                    $ref: '#/components/schemas/Events'
        "400":
          description: Invalid input or missing required fields
        "403":
          description: Unauthorized access, only admins can perform this action
        "404":
          description: Event not found
        "500":
          description: Server error
    
    delete:
      tags:
      - Event
      description: Deletes the event with the specified ID.
      parameters:
        - name: id
          in: path
          description: The unique ID of the event to delete
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Event successfully deleted."
                  event:
                    $ref: '#/components/schemas/Events'
        "400":
          description: Invalid event ID format
        "404":
          description: Event not found
        "500":
          description: Server error
...
  /pubPermissions:
    get:
      summary: Retrieve publication requests
      description: Returns all publication requests, optionally filtered by user_id.
      parameters:
        - name: user_id
          in: query
          required: false
          description: The ID of the user to filter requests by.
          schema:
            type: string
      responses:
        "200":
          description: Successful retrieval of publication requests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PubPermissions"
        "400":
          description: Invalid user_id format
        "404":
          description: No publication requests found
        "500":
          description: Server error

    post:
      summary: Create a new publication request
      description: Creates a new publication request with a user_id and description.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - user_id
                - Description
              properties:
                user_id:
                  type: string
                  description: The ID of the user creating the request.
                Description:
                  type: string
                  description: The description of the publication request.
      responses:
        "201":
          description: Publication request successfully created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  request:
                    $ref: "#/components/schemas/PubPermissions"
        "400":
          description: Invalid or missing user_id or description
        "500":
          description: Server error
  /pubPermissions/{id}:
    delete:
      summary: Delete a publication request
      description: Deletes a publication request by its ID.
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the publication request to delete.
          schema:
            type: string
      responses:
        "200":
          description: Publication request successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  request:
                    $ref: "#/components/schemas/PubPermissions"
        "400":
          description: Invalid request ID format
        "404":
          description: Publication request not found
        "500":
          description: Server error
...
components:
  schemas:
...
    Parks:
      type: object
      properties:
        id:
          type: string
          example: "60d21b4567d0d8992e610b98"
        name:
          type: string
          example: "Albere"
        x_coord:
          type: number
          example: 46.0709
        y_coord:
          type: number
          example: 11.1212
        rating:
          type: number
          example: 4.5
        description:
          type: string
          example: "A beautiful park with a fountain and running paths."
        categories:
          type: array
          items:
            type: string
            example: "Running, Pet-Friendly"
    Events:
      type: object
      required:
        - user_id
        - park_id
      properties:
        id:
          type: string
          description: Unique identifier for the event
          example: "60d21b4567d0d8992e610b98"
        user_id:
          type: string
          description: Reference to the ID of the user who created the event
          example: "60d21b4567d0d8992e610b97"
        park_id:
          type: string
          description: Reference to the ID of the park associated with the event
          example: "60d21b4567d0d8992e610b96"
        status:
          type: boolean
          description: Indicates whether the event is active
          default: false
          example: true
        description:
          type: string
          description: Additional details or notes about the event
          default: ""
          example: "Community park cleanup scheduled for Saturday morning."
...
    PubPermissions:
      type: object
      properties:
        id:
          type: string
          description: The unique identifier of the publication request.
        user_id:
          type: string
          description: The ID of the user who created the request.
        Description:
          type: string
          description: The description of the request.
    CategoryEnum:
      type: string
      enum:
        - Pet-Friendly
        - Sport
        - Running
        - Fountain
        - Nature
        - Playground
      example: "Running"
...